Split Places and Alerts into separate components
