+++ assets/css/screenplay-container.scss
@@ -25,4 +25,7 @@
       flex-grow: 1;
     }
   }
+  .page-content__hidden {
+    display: none;
+  }
 }
