+++ assets/js/components/App.tsx
@@ -15,14 +15,8 @@ class AppRoutes extends React.Component {
             path="/emergency-takeover"
             element={<OutfrontTakeoverTool />}
           ></Route>
-          <Route
-            path="/dashboard"
-            element={<Dashboard page="places" />}
-          ></Route>
-          <Route
-            path="/dashboard/alerts"
-            element={<Dashboard page="alerts" />}
-          ></Route>
+          <Route path="/dashboard" element={<Dashboard />}></Route>
+          <Route path="/dashboard/alerts" element={<Dashboard />}></Route>
         </Routes>
       </React.Suspense>
     );
