+++ assets/js/models/alert.ts
@@ -1,0 +1,6 @@
+export interface Alert {
+  id: string;
+  active_period: [Record<ActivePeriodLabel, string>];
+}
+
+type ActivePeriodLabel = "start" | "end";
